Create README.md
# NttCoin.sol
NttCoin contract

NttCoin protocol is a mixture of RFI tokenomics with the added function of auto-liquidity generating protocol. A fork from BEE with some added benefits.

First, the dev has minted the tokens - sent the balance to DxSale for fair launch - and then burned the remainder of the tokens. Dev had to partake in presale.

Second, utilized the DxSale protocol to foster trust in the community by auto-locking liquidity for 4 years.

Simple tokenomics and elegant design and production of token information has catered to the overall success of the project.

100% community owned and driven - no rug is possible. 

#NttCoin join us on our journey safely to the moon!